hass/python_scripts/dnull-mqtt/pyproject.toml
[tool.poetry]
name = "dnull-mqtt"
version = "0.1.0"
description = ""
authors = ["Bohdan Sukhomlinov <shellshock1953@gmail.com>"]
readme = "README.md"
packages = [{include = "dnull_mqtt"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
binance-connector = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
